RegistrationPeriodParseUtils 분리
